retoques

+++ src/main/java/entrega0/Producto.java
@@ -2,7 +2,7 @@ package entrega0;
 
 public class Producto extends CosaAcomprar {
 
-	public Producto(float precio) {
+	public Producto(double precio) {
 		this.precio = precio;
 		this.vendido = false;
 	}
