+++ src/main/java/entrega0/Servicio.java
@@ -2,7 +2,7 @@ package entrega0;
 
 public class Servicio extends CosaAcomprar {
 
-	public Servicio(float precio) {
+	public Servicio(double precio) {
 		this.precio = precio;
 		this.vendido = false;
 	}
